uasc: make response handlers type safe
This patch clarifies the naming of the messages received by a secure
channel and renames it from 'response' to MessageBody since that's
what it is. A MessageBody contains either a service request or response
depending on whether it sent by a client or a server.

As a consequence all client handlers have been typed to handle only
ua.Response objects instead of arbitrary interface{} values.

This is a preparation for the server code.